requirements.txt
minio~=7.1.8
pymongo~=4.1.1
Flask~=2.1.2
numpy~=1.22.3
nltk~=3.7
emoji~=1.7.0
num2words~=0.5.10
sklearn~=0.0
scikit-learn~=1.1.1
textblob~=0.17.1
regex~=2022.3.15
pyparsing~=3.0.8
python-dotenv